My_Medical_Chatbot/hf/requirements.txt
python-dotenv==1.0.1
pandas==2.1.4
langchain==0.2.6
langchain-openai==0.1.14
langchain-core==0.2.11
langchain-ibm==0.1.8
langchain-community==0.2.6
ibm-watson-machine-learning==1.0.359
ipykernel
notebook
urllib3 
requests==2.32.0
PyPDF2
python-docx
reportlab
openai
faiss-cpu
cryptography
pymysql
scikit-learn
gradio